use long instead of int to store the length of file

+++ app/src/main/java/com/aspsine/multithreaddownload/demo/ui/fragment/ListViewFragment.java
@@ -116,7 +116,7 @@ public class ListViewFragment extends Fragment implements OnItemClickListener<Ap
             }
 
             @Override
-            public void onConnected(int total, boolean isRangeSupport) {
+            public void onConnected(long total, boolean isRangeSupport) {
                 appInfo.setStatus(AppInfo.STATUS_DOWNLOADING);
                 if (isCurrentListViewItemVisible(position)) {
                     ListViewAdapter.ViewHolder holder = getViewHolder(position);
@@ -126,7 +126,7 @@ public class ListViewFragment extends Fragment implements OnItemClickListener<Ap
             }
 
             @Override
-            public void onProgress(int finished, int total, int progress) {
+            public void onProgress(long finished, long total, int progress) {
                 String downloadPerSize = getDownloadPerSize(finished, total);
                 appInfo.setProgress(progress);
                 appInfo.setDownloadPerSize(downloadPerSize);
@@ -196,7 +196,7 @@ public class ListViewFragment extends Fragment implements OnItemClickListener<Ap
         });
     }
 
-    private String getDownloadPerSize(int finished, int total) {
+    private String getDownloadPerSize(long finished, long total) {
         return DF.format((float) finished / (1024 * 1024)) + "M/" + DF.format((float) total / (1024 * 1024)) + "M";
     }
